Refactor Appointment commands
