+++ src/main/java/seedu/address/logic/commands/queue/UndoDequeueCommand.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.queue;
 
 import static java.util.Objects.requireNonNull;
