+++ src/main/java/seedu/address/logic/parser/AddressBookParser.java
@@ -6,23 +6,23 @@ import static seedu.address.commons.core.Messages.MESSAGE_UNKNOWN_COMMAND;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import seedu.address.logic.commands.AckAppCommand;
-import seedu.address.logic.commands.AddAppCommand;
-import seedu.address.logic.commands.AddConsultationRoomCommand;
-import seedu.address.logic.commands.AppointmentsCommand;
-import seedu.address.logic.commands.BreakCommand;
-import seedu.address.logic.commands.CancelAppCommand;
-import seedu.address.logic.commands.ChangeAppCommand;
-import seedu.address.logic.commands.DequeueCommand;
-import seedu.address.logic.commands.EnqueueCommand;
+import seedu.address.logic.commands.appointments.AckAppCommand;
+import seedu.address.logic.commands.appointments.AddAppCommand;
+import seedu.address.logic.commands.queue.AddConsultationRoomCommand;
+import seedu.address.logic.commands.appointments.AppointmentsCommand;
+import seedu.address.logic.commands.queue.BreakCommand;
+import seedu.address.logic.commands.appointments.CancelAppCommand;
+import seedu.address.logic.commands.appointments.ChangeAppCommand;
+import seedu.address.logic.commands.queue.DequeueCommand;
+import seedu.address.logic.commands.queue.EnqueueCommand;
 import seedu.address.logic.commands.ExitCommand;
 import seedu.address.logic.commands.HelpCommand;
-import seedu.address.logic.commands.MissAppCommand;
+import seedu.address.logic.commands.appointments.MissAppCommand;
 import seedu.address.logic.commands.NextCommand;
 import seedu.address.logic.commands.RedoCommand;
-import seedu.address.logic.commands.RemoveRoomCommand;
-import seedu.address.logic.commands.ResumeCommand;
-import seedu.address.logic.commands.SettleAppCommand;
+import seedu.address.logic.commands.queue.RemoveRoomCommand;
+import seedu.address.logic.commands.queue.ResumeCommand;
+import seedu.address.logic.commands.appointments.SettleAppCommand;
 import seedu.address.logic.commands.UndoCommand;
 import seedu.address.logic.commands.common.Command;
 import seedu.address.logic.commands.common.CommandHistory;
