+++ src/main/java/seedu/address/logic/parser/ChangeAppCommandTimingParser.java
@@ -10,7 +10,7 @@ import java.util.stream.Stream;
 
 import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.ChangeAppCommand;
+import seedu.address.logic.commands.appointments.ChangeAppCommand;
 import seedu.address.logic.commands.common.ReversibleActionPairCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
