+++ src/main/java/seedu/address/logic/parser/EnqueueCommandParser.java
@@ -4,8 +4,8 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.DequeueCommand;
-import seedu.address.logic.commands.EnqueueCommand;
+import seedu.address.logic.commands.queue.DequeueCommand;
+import seedu.address.logic.commands.queue.EnqueueCommand;
 import seedu.address.logic.commands.common.ReversibleActionPairCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReferenceId;
