+++ src/main/java/seedu/address/logic/parser/MissAppCommandParser.java
@@ -3,7 +3,7 @@ package seedu.address.logic.parser;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
-import seedu.address.logic.commands.MissAppCommand;
+import seedu.address.logic.commands.appointments.MissAppCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
