+++ src/main/java/seedu/address/logic/parser/ResumeCommandParser.java
@@ -4,9 +4,9 @@ import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT
 
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.index.Index;
-import seedu.address.logic.commands.BreakCommand;
+import seedu.address.logic.commands.queue.BreakCommand;
 import seedu.address.logic.commands.NextCommand;
-import seedu.address.logic.commands.ResumeCommand;
+import seedu.address.logic.commands.queue.ResumeCommand;
 import seedu.address.logic.commands.common.ReversibleActionPairCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
