+++ src/test/java/seedu/address/logic/commands/ResumeCommandTest.java
@@ -12,6 +12,7 @@ import java.util.Optional;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.address.logic.commands.queue.ResumeCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.queue.Room;
