+++ src/test/java/seedu/address/logic/commands/patients/UnregisterPatientCommandTest.java
@@ -11,7 +11,7 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.commons.core.Messages;
-import seedu.address.logic.commands.DequeueCommand;
+import seedu.address.logic.commands.queue.DequeueCommand;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.person.Person;
