+++ src/test/java/seedu/address/logic/parser/DequeueCommandParserTest.java
@@ -5,7 +5,7 @@ import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailur
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.DequeueCommand;
+import seedu.address.logic.commands.queue.DequeueCommand;
 import seedu.address.model.Model;
 import seedu.address.testutil.TestUtil;
