+++ src/test/java/seedu/address/logic/parser/EnqueueCommandParserTest.java
@@ -5,7 +5,7 @@ import static seedu.address.model.person.parameters.PersonReferenceId.MESSAGE_CO
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.EnqueueCommand;
+import seedu.address.logic.commands.queue.EnqueueCommand;
 
 public class EnqueueCommandParserTest {
     private EnqueueCommandParser parser = new EnqueueCommandParser();
